chore: update 'solid-router'
