+++ apps/solidjs-boilerplate/package.json
@@ -87,13 +87,13 @@
   },
   "dependencies": {
     "@sect/solid-hiding-header": "workspace:*",
-    "@solidjs/meta": "^0.28.6",
-    "@solidjs/router": "^0.8.3",
+    "@solidjs/meta": "^0.29.3",
+    "@solidjs/router": "^0.10.5",
     "@tanstack/solid-query": "^4.36.1",
     "escape-goat": "^4.0.0",
     "isomorphic-dompurify": "^1.9.0",
-    "solid-devtools": "^0.27.7",
-    "solid-js": "^1.8.0",
+    "solid-devtools": "^0.29.2",
+    "solid-js": "^1.8.7",
     "solid-toast": "^0.5.0",
     "striptags": "^3.2.0",
     "truncate": "^3.0.0"
