+++ packages/solid-hiding-header/package.json
@@ -88,7 +88,7 @@
     "@typescript-eslint/parser": "^6.7.5",
     "@vitest/coverage-v8": "^1.1.0",
     "@vue/test-utils": "^2.4.1",
-    "babel-preset-solid": "^1.8.0",
+    "babel-preset-solid": "^1.8.6",
     "eslint": "8.51.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-airbnb-typescript": "^17.1.0",
@@ -116,7 +116,7 @@
     "rollup": "^4.0.2",
     "rollup-preset-solid": "^2.0.1",
     "secretlint": "^7.0.7",
-    "solid-js": "^1.8.0",
+    "solid-js": "^1.8.7",
     "taze": "^0.11.3",
     "tsc-files": "^1.1.4",
     "typescript": "^5.2.2",
